backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic>=2.8.0
firebase-admin==5.2.0         
python-dotenv==1.0.0        
pytest==7.4.0    
email-validator>=2.0.0
joblib>=1.3.0
requests>=2.31.0
scikit-learn>=1.3.0
numpy>=1.24.0            
